ISSUE=12988 adding initial conversion of nddi_dbus to nddi_rabbitmq
